fbtft: Support sending commands at 1 MHz
We pass an additional parameter to write() and read() that indicates
whether we're doing a command or data operation.  If it's a command
operation, use a slower speeed.

This is done because some panels can operate at higher frequencies when
displaying data, but must operate at lower frequencies when configuring.

Signed-off-by: Sean Cross <xobs@kosagi.com>